Revival poller polls even if no players are online #70
